backend/requirements.txt
fastapi>=0.103.1
uvicorn>=0.23.2
pydantic>=2.3.0
sqlalchemy>=2.0.20
alembic>=1.12.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt==4.0.1
celery>=5.3.4
redis>=5.0.0
httpx>=0.24.1
openai>=1.12.0
python-dotenv>=1.0.0
psycopg2>=2.9.9
pydantic-settings>=2.0.0
email-validator>=2.0.0
asyncpg>=0.28.0
greenlet>=2.0.2
pyjwt>=2.10.1
numpy>=2.2.4
pgvector>=0.4.0
tiktoken>=0.9.0
pytz>=2024.1
scipy>=1.11.3
pillow>=10.1.0
opencv-python>=4.11.0.2
python-docx>=0.8.11
psutil>=5.9.5
redis[hiredis]>=5.0.0
pypdf>=4.2.0
flower>=2.0.0
gevent>=24.11.1